refactored and added more openai customization/optimization
